backend-requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
tweepy==4.14.0
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.6.0
